segy_trace_viewer/requirements.txt
numpy>=1.20.0
PyQt5>=5.15.0
PyOpenGL>=3.1.5
